try original location before our mirror
We added this because downloads were failing due to deno issues that only happened in CI, but now we use cURL in CI to work around this.

We should check the original, it may have been yanked and replaced for example.

Although longer term our protocol will work around that.