Ensure runtime HOMEs are set (#58)
Like really this is awful and should be part of useConfig() /cc @ABevier

Which would require us to *not* support `env.FOO` and instead only allow certain keys like `env.HOME` as the first and only one. I can't honestly think of anymore that we would allow off the bat.