Invoke test script directly
Invoking it with these args stops the command not found handler running. I don’t understand why it has that effect.

Also I just don’t know why we would invoke it this way. We clearly were surprised about that in the way that the shebang is defined as a pkgx shebang suggesting we also always want to use our own bash.